Remove encoder/decoder control codes which weren't in the 1.0 release from the legacy theora.h header. While the theora_control() implementation will pass on all the same control codes recognized by th_encode_ctl() and th_decode_ctl(), we want to encourage migration to the newer api, so we only declare them there.
Remove an obsolete limitation on user-supplied quant matrices.

Switch to an _old suffix on the theora.h version of the control code 
anchors to avoid confusion with the new api. I also tried to clean up 
the formatting but wasn't entirely successful; the documentation for 
theoraenc.h and theoradec.h looks better.