Add oc_state_mb_recon() and call it sometimes.  It doesn't really do a macroblock, though.  Just a quad of blocks.
Now to think about what information we really need, and where to find it, and whether or not it should be put somewhere more convenient.

Also need to accelerate some of the wide operations that have been opened up, and to recognise cases (most obviously 444) where chroma can do quads too.